Добавление зависимостей в s3mod файлы UICore и UICommon, решение проблем https://online.sbis.ru/opendoc.html?guid=ce57e942-2c02-46b8-a31a-beb0309f06b3
