+++ UI/Base.ts
@@ -11,6 +11,6 @@
  * @public
  * @author Шипин А.А.
  */
-export { Control, TemplateFunction } from 'UICore/Base';
-export { IControlOptions, IControlChildren } from 'UICommon/Base';
+export { Control } from 'UICore/Base';
+export { IControlOptions, IControlChildren, TemplateFunction } from 'UICommon/Base';
 export * from './_base/CommonLib';
