+++ UI/_async/Async.ts
@@ -3,8 +3,8 @@ import * as ModulesLoader from 'WasabyLoader/ModulesLoader';
 import * as Library from 'WasabyLoader/Library';
 import { IoC, constants } from 'Env/Env';
 import { descriptor } from 'Types/entity';
-import { Control, TemplateFunction } from 'UICore/Base';
-import { IControlOptions } from 'UICommon/Base';
+import { Control } from 'UICore/Base';
+import { IControlOptions, TemplateFunction } from 'UICommon/Base';
 import { headDataStore } from 'UI/Deps';
 import template = require('wml!UI/_async/Async');
