+++ UI/_base/Document.ts
@@ -1,6 +1,7 @@
 /// <amd-module name="UI/_base/Document" />
 
-import { Control, TemplateFunction } from 'UICore/Base';
+import { Control } from 'UICore/Base';
+import { TemplateFunction } from 'UICommon/Base';
 
 // tslint:disable-next-line:ban-ts-ignore
 // @ts-ignore
