+++ UICommon/UICommon.s3mod
@@ -3,7 +3,15 @@
 
   <description>Интерфейсы и общий функционал для UIReact и UIInferno</description>
 
-  <depends/>
+  <depends>
+    <module id="0df0fce3-53ba-47f8-987d-edf1fd078295" name="WS.Core"/>
+    <module id="636177e5-4d75-44fa-9173-977dc90ff111" name="Env"/>
+    <module id="323fd54f-1fab-42a6-8aea-6726d6392d53" name="Application"/>
+    <module id="686f183d-0bcc-4732-b4e4-6016c2a67a0f" name="Types"/>
+    <module id="37e402f1-2fb0-4b73-81c5-113cf94dbf1c" name="WasabyLoader"/>
+    <module id="7e8aabad-d60f-4c48-aa74-f29d78fd8be2" name="RequireJsLoader"/>
+    <module id="951e1c80-ee24-4022-8349-96123fae8026" name="Inferno"/>
+  </depends>
 
   <load_after/>
