+++ UICommon/_events/IEvents.ts
@@ -1,5 +1,5 @@
-import { IControlNode, IDOMEnvironment, IWasabyHTMLElement } from 'UICommon/interfaces';
-import {Control} from 'UICore/Base';
+import { ICommonControlNode as IControlNode, ICommonDOMEnvironment as IDOMEnvironment, IWasabyHTMLElement } from 'UICommon/interfaces';
+import { IControl } from 'UICommon/interfaces';
 
 type VoidFunction = () => void;
 
@@ -81,7 +81,7 @@ export interface IWasabyEvent {
     handler: Function;
     isControl: boolean;
     value: string;
-    viewController: Control;
+    viewController: IControl;
 }
 
 export interface IWasabyEventSystem {
