+++ UICommon/_executor/_Utils/Common.ts
@@ -10,13 +10,11 @@ import { constants, cookie } from 'Env/Env';
 
 import * as RequireHelper from './RequireHelper';
 
-import { ReactiveObserver } from 'UICore/Reactivity';
-import {
-   IControl,
-   IGeneratorNameObject
-} from '../_Markup/IGeneratorType';
+import { pauseReactive } from '../pauseReactive';
+import { IControl } from 'UICommon/interfaces';
+import { IGeneratorNameObject } from '../_Markup/IGeneratorType';
 import * as Attr from '../_Expressions/Attr';
-import { TemplateFunction } from 'UICore/Base';
+import { TemplateFunction } from 'UICommon/Base';
 
 var
    requireIfDefined = function requireIfDefined(tpl) {
@@ -187,7 +185,7 @@ export function plainMerge(inner, object, cloneFirst?) {
 
    if (cloneFirst) {
       // pause reactive behaviour of properties while merging
-      ReactiveObserver.pauseReactive(copyInner, () => {
+      pauseReactive(copyInner, () => {
          for (prop in inner) {
             if (inner.hasOwnProperty(prop)) {
                copyInner[prop] = inner[prop];
@@ -199,7 +197,7 @@ export function plainMerge(inner, object, cloneFirst?) {
    }
 
    // pause reactive behaviour of properties while merging
-   ReactiveObserver.pauseReactive(copyInner, () => {
+   pauseReactive(copyInner, () => {
       for (prop in object) {
          if (object.hasOwnProperty(prop)) {
             copyInner[prop] = object[prop];
