+++ UIInferno/UICore/UITest/Focus/Proxy.ts
@@ -1,4 +1,5 @@
-import {Control, TemplateFunction} from 'UICore/Base';
+import {Control} from 'UICore/Base';
+import { TemplateFunction } from 'UICommon/Base';
 
 // @ts-ignore
 import template = require('wml!UICore/UITest/Focus/Proxy');
