+++ UIInferno/UICore/_vdom/Synchronizer/resources/Hooks.ts
@@ -1,5 +1,9 @@
 import { EventUtils, isInvisibleNode } from 'UICommon/Events';
-import { IDOMEnvironment, IProperties, IWasabyHTMLElement, IControlNode, IEvent, TEventsObject, TControlId } from 'UICommon/interfaces';
+import { IProperties, IWasabyHTMLElement, IEvent, TEventsObject, TControlId } from 'UICommon/interfaces';
+import {
+    IDOMEnvironment,
+    IControlNode
+} from 'UICore/interfaces';
 import { constants } from 'Env/Env';
 import { Set } from 'Types/shim';
