+++ UIReact/UICore/ReactDemo/WasabyReactivity/Rec.ts
@@ -1,6 +1,7 @@
 import {RecordSet} from 'Types/collection';
 import {Model, Record} from 'Types/entity';
-import {Control, TemplateFunction} from 'UICore/Base';
+import {Control} from 'UICore/Base';
+import { TemplateFunction } from 'UICommon/Base';
 // @ts-ignore
 import * as template from 'wml!UICore/ReactDemo/WasabyReactivity/Rec';
