+++ UIReact/UICore/Reactivity.ts
@@ -1,8 +1,8 @@
+import { pauseReactive } from 'UICommon/Executor';
+
 export { makeObservable, useMakeObservable } from './_react/Reactivity/MakeObservable';
 export { withVersionObservable, getReactiveVersionsProp } from './_react/Reactivity/withVersionObservable';
 
 export const ReactiveObserver = {
-    pauseReactive: (_: any, func: Function) => {
-        func();
-    }
+    pauseReactive
 };
